モーダルをコンポネント化
