+++ back/app/controllers/api/v1/pages_controller.rb
@@ -5,7 +5,7 @@ class Api::V1::PagesController < ApplicationController
     Rails.logger.debug("Received params: #{params.inspect}")
     last_page_number = Page.where(book_id: page_params[:book_id]).maximum(:page_number) || 0
     page = Page.new(page_params.merge(page_number: last_page_number + 1))
-    page.background_color = page_params[:content][:backgroundColor]
+    page.background_color = page_params[:content][:background_color]
 
     if page.save
       render json: { page: page.as_json(include: [:page_characters, :page_elements]) }, status: :created
@@ -30,15 +30,12 @@ class Api::V1::PagesController < ApplicationController
     params.require(:page).permit(
       :book_id,
       :page_number,
+      :background_color,
       content: [
-        :title,
-        :author,
-        :tags,
         :backgroundColor,
-        :visibility,
         texts: [
           :text,
-          :fontSize,
+          :font_size,
           :color,
           :x,
           :y,
@@ -51,27 +48,8 @@ class Api::V1::PagesController < ApplicationController
           :x,
           :y,
           :width,
-          :height,
-          :rotation,
-          :scaleX,
-          :scaleY
+          :height
         ]
-      ],
-      page_characters_attributes: [
-        :character_type,
-        :simple_path,
-        :body_path,
-        :hair_path,
-        :eye_path,
-        :mouth_path,
-        :hand_path,
-        :foot_path,
-        :outfit_path,
-        :position_x,
-        :position_y,
-        :rotation,
-        :scale_x,
-        :scale_y
       ],
       page_elements_attributes: [
         :element_type,
@@ -83,10 +61,7 @@ class Api::V1::PagesController < ApplicationController
           :position_y,
           :src,
           :width,
-          :height,
-          :rotation,
-          :scaleX,
-          :scaleY
+          :height
         ]
       ]
     )
